:recycle: move data processing to \data_process.py
More specifically, move `soup_to_weekly_taskboards( )` and `update_taskboards_with_stuefordeling( )` to another file, which I have elected to call \data_process.py
